requirements.txt
moviepy==2.2.1
numpy==1.25.0
onnx==1.18.0
onnxruntime==1.19.2
openai-whisper==20250625
opencv-python==4.11.0.86
pydub==0.25.1
scipy==1.11.4
TTS==0.22.0
torch==2.7.1
tqdm==4.67.1
ultralytics==8.3.159
